requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
greenlet==3.3.2
gunicorn==25.3.0
h11==0.16.0
idna==3.11
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
psycopg2-binary
python-dotenv
python-multipart
supabase